File up down

+++ basic-mysql/controllers/admin.js
@@ -1,3 +1,6 @@
+
+
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
 const Product = require('../models/product');
 
@@ -14,9 +17,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: []
+    });
+  }
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()){
@@ -35,6 +54,8 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title, 
@@ -58,7 +79,7 @@ exports.postAddProduct = (req, res, next) => {
       //   hasError: true,
       //   product: {
       //     title: title,
-      //     imageUrl: imageUrl,
+      //     image: image,
       //     price: price,
       //     description: description
       //   },
@@ -103,7 +124,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedPrice = req.body.price;
   const updatedDesc = req.body.description;
 
@@ -117,7 +138,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -133,10 +153,12 @@ exports.postEditProduct = (req, res, next) => {
     }
     product.title = updatedTitle;
     product.price = updatedPrice;
-    product.imageUrl = updatedImageUrl;
     product.description = updatedDesc;
-    product
-      .save().then(result => {
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
+    return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
       });
@@ -169,14 +191,20 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log('DESTROY PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error = new Error(err)
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      return next(new Error('Product notr found'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: prodId, userId: req.user._id })
+  })
+  .then(() => {
+    console.log('DESTROY PRODUCT');
+    res.redirect('/admin/products');
+  })
+  .catch(err => {
+    const error = new Error(err)
+    error.httpStatusCode = 500;
+    return next(error);
+  });
 };
